refactor: lib directory
