+++ frontend/package.json
@@ -34,6 +34,7 @@
   "devDependencies": {
     "@eslint/js": "^9.36.0",
     "@hey-api/openapi-ts": "^0.88",
+    "@types/lodash": "^4.17.23",
     "@types/luxon": "^3.7.1",
     "@types/node": "^24.3.1",
     "@types/react": "^19.0.8",
