+++ frontend/src/components/index.tsx
@@ -1,5 +1,6 @@
 export { default as Button } from './Button';
 export { default as Canvas } from './Canvas';
+export { default as ContextMenu } from './ContextMenu';
 export { default as Dialog } from './Dialog';
 export { default as Dropdown } from './Dropdown';
 export { default as ErrorBanner } from './ErrorBanner';
