+++ frontend/src/containers/DateNav.jsx
@@ -1,9 +1,8 @@
 import { Button } from '@components';
 import { useDialog } from '@features/Dialogs';
+import { dateToDateString } from '@lib/utils';
 import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
-
-import { dateToDateString } from '/src/utils';
 
 const DateNav = ({ onChange, skipBy = 1 }) => {
   const [searchParams, setSearchParams] = useSearchParams();
