--- frontend/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useLocalStorage } from './useLocalStorage';
-export { default as useKeyDown } from './useKeyDown';
