+++ frontend/vite.config.js
@@ -9,7 +9,7 @@ export default ({ mode }) => {
         '@containers': '/src/containers',
         '@components': '/src/components',
         '@features': '/src/features',
-        '@hooks': '/src/hooks',
+        '@lib': '/src/lib',
         '@types': '/src/types',
         '@': '/src',
       },
